+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+- NA
+
+## [0.3.0] - 2025-03-23
+
 - Added [Sponge-Property] function
 - Added [Sponge-Script] function
 - Added [Split-HereString] function
+- Added [Set-DotEnv] Support for GnuPG decryption
 
 ## [0.2.0] - 2025-03-09
 
@@ -236,6 +241,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[unreleased]: https://github.com/btklab/pwsh-sketches/compare/0.2.0..HEAD
+[unreleased]: https://github.com/btklab/pwsh-sketches/compare/0.3.0..HEAD
+[0.3.0]: https://github.com/btklab/pwsh-sketches/releases/tag/0.3.0
 [0.2.0]: https://github.com/btklab/pwsh-sketches/releases/tag/0.2.0
 [0.1.0]: https://github.com/btklab/pwsh-sketches/releases/tag/0.1.0